Ensures debug keystore creation in CI
Addresses an issue where the debug keystore might be missing during CI builds, specifically in manual build mode.

This commit adds logic to check for the existence of `debug.keystore`
and generates it if it's not found. This ensures consistent builds
even when the keystore is not explicitly present in the repository.
Also, prevent dependabot to create debug keystore.

Fixes missing debug keystore for CI builds.